@@ -21,16 +21,14 @@
         "@knownout/lib": "^0.0.8",
         "@testing-library/react": "^12.1.4",
         "@types/jest": "^27.4.1",
-        "@types/react": "^17.0.41",
-        "@types/react-dom": "^17.0.14",
         "babel-loader": "^8.2.3",
         "css-loader": "^6.7.1",
         "html-webpack-plugin": "^5.5.0",
         "jest": "^27.5.1",
         "mini-css-extract-plugin": "^2.6.0",
         "postcss-loader": "^6.2.1",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0",
         "recoil": "^0.7.3-alpha.1",
         "sass": "^1.49.9",
         "sass-loader": "^12.6.0",
